+++ backend/xemphim/src/main/java/com/example/xemphim/Repository/GenreRepository.java
@@ -5,6 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Collection;
+import java.util.Set;
+
 @Repository
-public interface GenreRepository  extends JpaRepository<Genre, Integer> {
+public interface GenreRepository  extends JpaRepository<Genre, Long> {
+
 }
